agno-agi/beginner/deep-research-agent/requirements.txt
agno
openai
psycopg
firecrawl
streamlit
sqlalchemy
python-dotenv
fastapi[standard]

